Refine CLI tests support helpers
